uslovi izvrsavanja

+++ airbnb/react-typescript-api-call/src/components/ReservationGuest.tsx
@@ -6,13 +6,22 @@ import ReservationService from "../services/ReservationService";
 const ReservationGuest = ({reservation}) => {
 
     const remove = (e:any) => {
-        ReservationService.delete(reservation.id).then((response) => {
-            console.log(response);
-        })
-            .catch((error) => {
-                console.log(error);
-            });
-        window.location.reload();
+        const newDate = new Date(reservation.startDate)
+        const now = new Date()
+        const diff = Math.abs(newDate.getTime() - now.getTime())
+        const diffDays = Math.ceil(diff / (1000 * 3600 * 24))
+        if (diffDays > 1) {
+            ReservationService.delete(reservation.id).then((response) => {
+                console.log(response);
+            })
+                .catch((error) => {
+                    console.log(error);
+                });
+            window.location.reload();
+        }
+        else{
+            alert("Less than 1 day till reservation begin")
+        }
     }
     return(
         <TableRow key={reservation.id}>
@@ -30,6 +39,9 @@ const ReservationGuest = ({reservation}) => {
             </TableCell>
             <TableCell align={"center"}>
                 <div >{reservation.number}</div>
+            </TableCell>
+            <TableCell align={"center"}>
+                <div >{reservation.status}</div>
             </TableCell>
             <TableCell align={"center"}>
                 <div><button onClick={remove}>Delete</button></div>
